Added security features to JournalEntryController APIs

+++ src/main/java/co/in/HSBC/journalapp/services/JournalEntryService.java
@@ -44,11 +44,21 @@ public class JournalEntryService {
         journalRepository.save(journalEntry);
     }
 
-    public void removeEntry(ObjectId id, String username) {
-        User user = userService.findUserByUsername(username);
-        user.getJournalEntries().removeIf(x -> x.getId().equals(id));
-        userService.saveUser(user);
-        journalRepository.deleteById(id);
+    @Transactional
+    public boolean removeEntry(ObjectId id, String username) {
+        boolean removed = false;
+        try {
+            User user = userService.findUserByUsername(username);
+            removed = user.getJournalEntries().removeIf(x -> x.getId().equals(id));
+            if (removed) {
+                userService.saveUser(user);
+                journalRepository.deleteById(id);
+            }
+        } catch (Exception e) {
+            System.out.println(e);
+            throw new RuntimeException("An error occurred while deleting the entry. " + e);
+        }
+        return removed;
     }
 
     public Optional<JournalEntry> findById(ObjectId id) {
